Remove unused classes, removed white spaces, moved files to the correct folder
