+++ MercadoPagoSDK/MercadoPagoSDK/Core/Extensions/Date+Additions.swift
@@ -8,15 +8,6 @@
 
 import Foundation
 internal extension Date {
-    func getCurrentMillis() -> Int64 {
-        return Int64(self.timeIntervalSince1970 * 1000)
-    }
-
-    func from(millis: Int64) -> Date {
-        let timeInterval: TimeInterval = Double(millis) / 1000
-        return Date(timeIntervalSince1970: timeInterval)
-    }
-
     func stringDate() -> String {
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "yyyy-MM-dd HH:mm:ss"
