+++ MercadoPagoSDK/MercadoPagoSDK/Core/Extensions/String+Additions.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-
 internal extension String {
 
     static let NON_BREAKING_LINE_SPACE = "\u{00a0}"
@@ -85,5 +84,13 @@ internal extension String {
 
     var isNumber: Bool {
         return rangeOfCharacter(from: CharacterSet.decimalDigits.inverted) == nil
+    }
+    
+    func insert(_ string: String, ind: Int) -> String {
+        return  String(self.prefix(ind)) + string + String(self.suffix(self.count - ind))
+    }
+    
+    var isNotEmpty: Bool {
+        return !isEmpty
     }
 }
