--- MercadoPagoSDK/MercadoPagoSDK/UI/Base/PXMonospaceLabel.swift
@@ -1,13 +1,0 @@
-//
-//  PXMonospaceLabel.swift
-//  MercadoPagoSDK
-//
-//  Created by Demian Tejo on 29/1/18.
-//  Copyright © 2018 MercadoPago. All rights reserved.
-//
-
-import UIKit
-
-class PXMonospaceLabel: UILabel {
-
-}
